challenge 2/requirements.txt
numpy
pandas
scikit-learn
Pillow
tqdm
tensorflow
opencv-python
